pyproject.toml
[project]
name = "tokenbowl-mcp"
version = "0.1.0"
description = "MCP server for Token Bowl fantasy football league via Sleeper API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.11.3",
    "httpx>=0.28.1",
    "redis>=5.0.0",
    "python-dotenv>=1.0.0",
    "logfire[httpx]>=4.4.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-vcr>=1.0.2",
    "vcrpy>=6.0.0",
    "fakeredis>=2.20.0",
]

dev = [
    "ruff>=0.12.0",
    "pre-commit>=3.5.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.12",
    "vcrpy>=7.0.0",
]
